Refactor Theo folders
