+++ src/app/theo/annonces/[id]/(detailAnnonce)/layout.tsx
