+++ src/app/theo/newsletter/confirm/loading.tsx
