+++ src/app/theo/newsletter/delete/page.tsx
